batch/requirements.txt
google-cloud-pubsub
kubernetes
pyyaml
